Add files via upload
ic-Tac-Toe: Classic Game Reinvented is a modern digital adaptation of the timeless pencil-and-paper game. Offering a sleek and intuitive user interface, this project aims to provide an engaging and immersive gaming experience for players of all ages. Built with versatility in mind, the project can be deployed across multiple platforms, including web browsers, mobile devices, and desktop applications.